requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
markdown==3.5.1
python-dotenv==1.0.0
